# internal/apidocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
        },
        "description": "API for managing AIG Custom Platform AI security scanning tasks. Based on Tencent Zhuque Lab AI-Infra-Guard (https://github.com/Tencent/AI-Infra-Guard).",
        "title": "AIG Custom Platform 任务API",
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/platform/admin/users": {
            "get": {
                "description": "Lists local accounts in a paginated safe envelope. Administrator only; credential material is never returned. page defaults to 1, page_size defaults to 20, values over 100 are capped at 100, and page cannot exceed 1000.",
                "parameters": [
                    {
                        "default": 1,
                        "in": "query",
                        "maximum": 1000,
                        "minimum": 1,
                        "name": "page",
                        "type": "integer"
                    },
                    {
                        "default": 20,
                        "in": "query",
                        "maximum": 100,
                        "minimum": 1,
                        "name": "page_size",
                        "type": "integer"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "platform-admin"
                ],
                "summary": "List managed users",
                "responses": {
                    "200": {
                        "description": "User metadata without password hashes.",
                        "schema": {
                            "$ref": "#/definitions/admin.UserListResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid page or page_size."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "403": {
                        "description": "Administrator role and completed password change required."
                    },
                    "500": {
                        "description": "Account query failed; internal details are not exposed."
                    }
                }
            },
            "post": {
                "description": "Creates a local account with an administrator-assigned role and records an audit event. Passwords and hashes are never returned.",
                "tags": [
                    "platform-admin"
                ],
                "summary": "Create managed user",
                "responses": {
                    "201": {
                        "description": "User created and audited."
                    },
                    "403": {
                        "description": "Administrator role required."
                    }
                }
            }
        },
        "/api/v1/platform/admin/users/{userID}/role": {
            "put": {
                "description": "Assigns admin, user, or auditor and records an audit event. Administrator only.",
                "tags": [
                    "platform-admin"
                ],
                "summary": "Assign user role",
                "parameters": [
                    {
                        "in": "path",
                        "name": "userID",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Role assigned and audited."
                    },
                    "403": {
                        "description": "Administrator role required."
                    }
                }
            }
        },
        "/api/v1/platform/admin/users/{userID}/active": {
            "put": {
                "description": "Enables or disables an account, revokes sessions on disable, and records an audit event. Administrator only.",
                "tags": [
                    "platform-admin"
                ],
                "summary": "Set account active state",
                "parameters": [
                    {
                        "in": "path",
                        "name": "userID",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Account state changed and audited."
                    },
                    "403": {
                        "description": "Administrator role required."
                    }
                }
            }
        },
        "/api/v1/platform/admin/users/{userID}/password-reset": {
            "post": {
                "description": "Requests a password reset through the configured out-of-band channel and records an audit event. The HTTP response never contains a reset token.",
                "tags": [
                    "platform-admin"
                ],
                "summary": "Request managed password reset",
                "parameters": [
                    {
                        "in": "path",
                        "name": "userID",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Reset requested and audited; no token is returned."
                    },
                    "403": {
                        "description": "Administrator role required."
                    }
                }
            }
        },
        "/api/v1/platform/admin/audit-events": {
            "get": {
                "description": "Queries append-only recursively sanitized audit events in a paginated envelope. Governed mutations correlate durable pending and completion evidence by request_id. Administrators and auditors may read; ordinary users are denied. page defaults to 1, page_size defaults to 20, values over 100 are capped at 100, and page cannot exceed 1000.",
                "parameters": [
                    {
                        "in": "query",
                        "name": "action",
                        "type": "string"
                    },
                    {
                        "in": "query",
                        "name": "actor_user_id",
                        "type": "string"
                    },
                    {
                        "in": "query",
                        "name": "resource_type",
                        "type": "string"
                    },
                    {
                        "in": "query",
                        "name": "resource_id",
                        "type": "string"
                    },
                    {
                        "default": 1,
                        "in": "query",
                        "maximum": 1000,
                        "minimum": 1,
                        "name": "page",
                        "type": "integer"
                    },
                    {
                        "default": 20,
                        "in": "query",
                        "maximum": 100,
                        "minimum": 1,
                        "name": "page_size",
                        "type": "integer"
                    }
                ],
                "tags": [
                    "audit"
                ],
                "summary": "Query audit events",
                "responses": {
                    "200": {
                        "description": "Sanitized audit event page.",
                        "schema": {
                            "$ref": "#/definitions/admin.AuditListResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid page or page_size."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "403": {
                        "description": "Administrator or auditor role and completed password change required."
                    },
                    "500": {
                        "description": "Audit query failed; internal details are not exposed."
                    }
                }
            }
        },
        "/api/v1/platform/admin/audit-events/prepared/{requestID}/finalize": {
            "post": {
                "description": "Administrator-only recovery for a prepared file-backed mutation after an operator verifies its real outcome. It promotes the existing request to the same stable outbox and event identity without repeating the file mutation. Retrying the same outcome is idempotent; a conflicting outcome returns 409. The request_id is obtained from the audit-event query or the safe X-Audit-Request-ID header returned when a successful legacy mutation needs recovery.",
                "tags": [
                    "audit"
                ],
                "summary": "Finalize a prepared audit completion",
                "parameters": [
                    {
                        "in": "path",
                        "name": "requestID",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "required": [
                                "outcome"
                            ],
                            "properties": {
                                "outcome": {
                                    "type": "string",
                                    "enum": [
                                        "success",
                                        "failure"
                                    ]
                                },
                                "metadata": {
                                    "type": "object",
                                    "additionalProperties": true
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "The prepared completion is durably ready for reconciliation."
                    },
                    "400": {
                        "description": "Invalid outcome or request body."
                    },
                    "403": {
                        "description": "Administrator role required."
                    },
                    "404": {
                        "description": "Prepared request not found."
                    },
                    "409": {
                        "description": "The request was already finalized with a different outcome."
                    }
                }
            }
        },
        "/api/v1/platform/admin/audit-events/reconcile": {
            "post": {
                "description": "Retries durable audit completion outbox entries whose final append previously failed. Delivery uses a stable event identifier, so repeated reconciliation does not duplicate audit events. Administrator only.",
                "tags": [
                    "audit"
                ],
                "summary": "Reconcile pending audit completions",
                "parameters": [
                    {
                        "in": "query",
                        "name": "limit",
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Number of completion events reconciled."
                    },
                    "403": {
                        "description": "Administrator role required."
                    }
                }
            }
        },
...
    },
    "definitions": {
        "admin.AuditListResponse": {
            "description": "Paginated sanitized audit-event envelope.",
            "properties": {
                "items": {
                    "items": {
                        "$ref": "#/definitions/audit.Event"
                    },
                    "type": "array"
                },
                "page": {
                    "minimum": 1,
                    "type": "integer"
                },
                "page_size": {
                    "maximum": 100,
                    "minimum": 1,
                    "type": "integer"
                },
                "total": {
                    "format": "int64",
                    "minimum": 0,
                    "type": "integer"
                }
            },
            "required": [
                "items",
                "total",
                "page",
                "page_size"
            ],
            "type": "object"
        },
        "admin.UserListResponse": {
            "description": "Administrator-only paginated local-account envelope.",
            "properties": {
                "items": {
                    "items": {
                        "$ref": "#/definitions/admin.UserResponse"
                    },
                    "type": "array"
                },
                "page": {
                    "minimum": 1,
                    "type": "integer"
                },
                "page_size": {
                    "maximum": 100,
                    "minimum": 1,
                    "type": "integer"
                },
                "total": {
                    "format": "int64",
                    "minimum": 0,
                    "type": "integer"
                }
            },
            "required": [
                "items",
                "total",
                "page",
                "page_size"
            ],
            "type": "object"
        },
        "admin.UserResponse": {
            "description": "Safe local-account metadata; credentials and credential hashes are not representable.",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "created_at": {
                    "format": "date-time",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "must_change_password": {
                    "type": "boolean"
                },
                "role": {
                    "enum": [
                        "admin",
                        "user",
                        "auditor"
                    ],
                    "type": "string"
                },
                "updated_at": {
                    "format": "date-time",
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            },
            "required": [
                "id",
                "username",
                "role",
                "active",
                "must_change_password",
                "created_at",
                "updated_at"
            ],
            "type": "object"
        },
        "audit.Event": {
            "description": "Append-only event with recursively sanitized metadata. Secret-like metadata keys are replaced by a redaction marker before persistence and response.",
            "properties": {
                "action": {
                    "description": "Governance action, including attachment.download_authorized for an approved administrator cross-owner download.",
                    "type": "string"
                },
                "actor_role": {
                    "enum": [
                        "admin",
                        "user",
                        "auditor"
                    ],
                    "type": "string"
                },
                "actor_user_id": {
                    "type": "string"
                },
                "actor_username": {
                    "type": "string"
                },
                "client_ip": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "metadata": {
                    "additionalProperties": true,
                    "description": "Sanitized JSON metadata only.",
                    "type": "object"
                },
                "occurred_at": {
                    "format": "date-time",
                    "type": "string"
                },
                "outcome": {
                    "enum": [
                        "pending",
                        "success",
                        "failure"
                    ],
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                },
                "resource_id": {
                    "type": "string"
                },
                "resource_type": {
                    "type": "string"
                }
            },
            "required": [
                "id",
                "occurred_at",
                "action",
                "outcome",
                "metadata"
            ],
            "type": "object"
        },
...
    }
}